dbparser: extract PDBRule into a separate module
