+++ modules/dbparser/Makefile.am
@@ -6,6 +6,8 @@ modules_dbparser_libsyslog_ng_patterndb_la_SOURCES	=	\
 	modules/dbparser/patterndb.c				\
 	modules/dbparser/patterndb.h				\
 	modules/dbparser/patterndb-int.h			\
+	modules/dbparser/pdb-rule.c				\
+	modules/dbparser/pdb-rule.h				\
 	modules/dbparser/pdb-action.c				\
 	modules/dbparser/pdb-action.h				\
 	modules/dbparser/pdb-program.c				\
